lab5 things

+++ lab2/C2/C2/IBuyTicket.cs
@@ -8,6 +8,6 @@ namespace C2
 {
     interface IBuyTicket
     {
-        abstract int BuyTicket();
+        abstract float BuyTicket();
     }
 }
